Update sudoku_solver.py
Implemented methods to check whether a number is only possible within one third of a row/column and eliminating them from the same box outside of that row/column. Also added function that allows the solver to check whether a pair of numbers are only possible in one pair of cells in a row/column/box and removing all other possibilities from those cells. This allows it to solve some hard sudokus but not all (The Sudoku featured as HARD is the sudoku featured in this Cracking the Cryptic video https://www.youtube.com/watch?v=4GVyBiFUNws